Refactor Database implementations
Migrate DB export out of client-side runningfile (Appcontext)

Update all databse imports
