+++ utils/AppContext.tsx
@@ -1,12 +1,7 @@
 /** @format */
 
-import React from "react";
-import { MongoClient } from "mongodb";
+import React, { useEffect, useState, createContext} from "react";
 import useLocalStorage from "use-local-storage";
-
-// Database connection
-const DB_CLIENT = new MongoClient(process.env.DB_CONN_STRING!);
-export const DB = DB_CLIENT.db(process.env.DB_NAME);
 
 export enum WindowWidth {
   SMALL,
@@ -20,7 +15,7 @@ interface AppContextProps {
   mobile: boolean;
 }
 
-export const AppContext = React.createContext<AppContextProps>({
+export const AppContext = createContext<AppContextProps>({
   windowWidth: WindowWidth.LARGE,
   theme: "dark",
   setTheme: () => {},
@@ -29,24 +24,24 @@ export const AppContext = React.createContext<AppContextProps>({
 
 export const AppContextProvider = ({ children }) => {
   const [theme, setTheme] = useLocalStorage("theme", "dark");
-  const [windowWidth, setWindowWidth] = React.useState(WindowWidth.LARGE);
-  const [mobile, setMobile] = React.useState(false);
+  const [windowWidth, setWindowWidth] = useState(WindowWidth.LARGE);
+  const [mobile, setMobile] = useState(false);
 
   // Grab User theme preference on initial app load
-  React.useEffect(() => {
+  useEffect(() => {
     const matchMedia = window.matchMedia("(prefers-color-scheme: dark)");
     setTheme(matchMedia.matches ? "dark" : "light");
   }, [setTheme]);
 
   // Theme change handling
-  React.useEffect(() => {
+  useEffect(() => {
     document.body.dataset.theme = theme;
   }, [theme]);
   const switchMode = () => {
     setTheme(theme === "dark" ? "light" : "dark");
   };
 
-  React.useEffect(() => {
+  useEffect(() => {
     // Check if viewing on mobile or desktop
     if (
       /Android|webOS|iPhone|iPad|iPod|BlackBerry/i.test(navigator.userAgent)
